visual-saas/backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
google-generativeai==0.8.5
Pillow==12.1.0
beautifulsoup4==4.14.3
requests==2.32.5
